+++ plugins/item.lua
@@ -92,10 +92,8 @@ Item.TYPE = {
 -- ========== Static Methods ==========
 
 Item.new = function(namespace, identifier, no_log)
-    if Item.find(namespace, identifier) then
-        log.error("Item already exists", 2)
-        return nil
-    end
+    local item = Item.find(namespace, identifier)
+    if item then return item end
 
     -- Create item
     local item = gm.item_create(