+++ plugins/object.lua
@@ -53,10 +53,8 @@ Object.CUSTOM_START = 800
 -- ========== Static Methods ==========
 
 Object.new = function(namespace, identifier, parent)
-    if Object.find(namespace, identifier) then
-        log.error("Object already exists", 2)
-        return nil
-    end
+    local obj = Object.find(namespace, identifier)
+    if obj then return obj end
 
     local obj = gm.object_add_w(namespace, identifier, Wrap.unwrap(parent))
     return Object.wrap(obj)